Keypad script system refactor (MP not tested)
This commit changes how the scripts of keypads works. Including:
- Replacement functions that replaces redundant calls and #define, namely fnc_buttonExecExpression and fnc_addActionCardReader (Thanks @ampersand38 !)
- New functions for validating memory point to see if such memory point exists. fnc_memoryPointValidation.
- New function that allows to have delayed activation with audio, much like the Lockroom from SCP:CB. fnc_buttonTriggerDelay. WIP
- New audio for the Lockroom alert used in fnc_buttonTriggerDelay.
- Moved #define of redundant attributes to config_macros_attributes.hpp from config.cpp.
- Minor adjustment to icons of Hallway_Corner_1 and Room_Endroom.
