+++ addons/room_things/CfgFunctions.hpp
@@ -5,6 +5,10 @@ class CfgFunctions {
             class buttonSound { file = QPATHTOF(functions\fnc_buttonSound.sqf); };
 			class clearanceLevelCheck { file = QPATHTOF(functions\fnc_clearanceLevelCheck.sqf); };
 			class buttonTrigger { file = QPATHTOF(functions\fnc_buttonTrigger.sqf); };
+			class addActionCardReader { file = QPATHTOF(functions\fnc_addActionCardReader.sqf); };
+			class memoryPointValidation { file = QPATHTOF(functions\fnc_memoryPointValidation.sqf); };
+			class buttonExecExpression { file = QPATHTOF(functions\fnc_buttonExecExpression.sqf); };
+			class buttonTriggerDelay { file = QPATHTOF(functions\fnc_buttonTriggerDelay.sqf); };
 		};
 	};
 };
