+++ addons/room_things/config_macros_attributes.hpp
@@ -1,227 +1,107 @@
-#define DOOR_GLASS_HITPOINT(glassID,arm,rad) \
-			class Glass_##glassID##_hitpoint \
-			{ \
-				armor = arm; \
-				material = -1; \
-				name = Glass_##glassID; \
-				visual = Glass_##glassID##_hide; \
-				passThrough = 0; \
-				radius = rad; \
-				convexComponent = Glass_##glassID##_hide; \
-				class DestructionEffects \
-				{ \
-					class BrokenGlass1 \
-					{ \
-						simulation = "particles"; \
-						type = "BrokenGlass1ND"; \
-						position = Glass_##glassID##_effects; \
-						intensity = 0.15000001; \
-						interval = 1; \
-						lifeTime = 0.05; \
-					}; \
-					class BrokenGlass2: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass2ND"; \
-					}; \
-					class BrokenGlass3: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass3ND"; \
-					}; \
-					class BrokenGlass4: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass4ND"; \
-					}; \
-					class BrokenGlass5: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass5ND"; \
-					}; \
-					class BrokenGlass6: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass6ND"; \
-					}; \
-					class BrokenGlass7: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass7ND"; \
-					}; \
-					class BrokenGlass1S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass1SD"; \
-					}; \
-					class BrokenGlass2S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass2SD"; \
-					}; \
-					class BrokenGlass3S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass3SD"; \
-					}; \
-					class BrokenGlass4S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass4SD"; \
-					}; \
-					class BrokenGlass5S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass5SD"; \
-					}; \
-					class BrokenGlass6S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass6SD"; \
-					}; \
-					class BrokenGlass7S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass7SD"; \
-					}; \
-				}; \
-			};
+#define ATTRIBUTES_DOOR_HINGE_STATE \
+class DoorHingeState \
+{ \
+	displayName     = CSTRING(DoorHingeState_displayname); \
+	tooltip         = CSTRING(DoorHingeState_tooltip); \
+	property        = "DoorHingeState"; \
+	control         = "Combo"; \
+	expression      = "_this animateSource ['Door_1_Source', _value, true];"; \
+	defaultValue    = 0; \
+	typeName        = "NUMBER"; \
+	class Values { \
+		class Close { \
+			name    = CSTRING(DoorHingeState_Close); \
+			tooltip = CSTRING(DoorHingeState_Close_tooltip); \
+			value   = 0; \
+			picture = QPATHTOF(data\Keycard\UI\gear_Keycard_C1_CA.paa); \
+		}; \
+		class Open { \
+			name    = CSTRING(DoorHingeState_Open); \
+			tooltip = CSTRING(DoorHingeState_Open_tooltip); \
+			value   = 1; \
+			picture = QPATHTOF(data\Keycard\UI\gear_Keycard_C2_CA.paa); \
+		}; \
+	}; \
+};
 
-#define BIG_GLASS_HITPOINT(glassID,arm,rad) \
-			class Glass_##glassID##_hitpoint \
-			{ \
-				armor = arm; \
-				material = -1; \
-				name = Glass_##glassID; \
-				visual = Glass_##glassID##_hide; \
-				passThrough = 0; \
-				radius = rad; \
-				convexComponent = Glass_##glassID##_hide; \
-				class DestructionEffects \
-				{ \
-					class BrokenGlass1 \
-					{ \
-						simulation = "particles"; \
-						type = "BrokenGlass1NB"; \
-						position = Glass_##glassID##_effects; \
-						intensity = 0.15000001; \
-						interval = 1; \
-						lifeTime = 0.05; \
-					}; \
-					class BrokenGlass2: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass2NB"; \
-					}; \
-					class BrokenGlass3: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass3NB"; \
-					}; \
-					class BrokenGlass4: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass4NB"; \
-					}; \
-					class BrokenGlass5: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass5NB"; \
-					}; \
-					class BrokenGlass6: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass6NB"; \
-					}; \
-					class BrokenGlass7: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass7NB"; \
-					}; \
-					class BrokenGlass1S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass1SB"; \
-					}; \
-					class BrokenGlass2S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass2SB"; \
-					}; \
-					class BrokenGlass3S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass3SB"; \
-					}; \
-					class BrokenGlass4S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass4SB"; \
-					}; \
-					class BrokenGlass5S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass5SB"; \
-					}; \
-					class BrokenGlass6S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass6SB"; \
-					}; \
-					class BrokenGlass7S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass7SB"; \
-					}; \
-				}; \
-			};
+#define ATTRIBUTES_DOOR_LOCK_STATE \
+class DoorLockState \
+{ \
+	displayName     = CSTRING(DoorLockState_displayname); \
+	tooltip         = CSTRING(DoorLockState_tooltip); \
+	property        = "DoorLockState"; \
+	control         = "Combo"; \
+	expression      = "_this setVariable ['SCP_CB_Operetable_DoorLockState', _value];"; \
+	defaultValue    = 0; \
+	typeName        = "NUMBER"; \
+	class Values { \
+		class Level1 { \
+			name    = CSTRING(DoorState_level1); \
+			tooltip = CSTRING(DoorState_level1_tooltip); \
+			value   = 1; \
+			picture = QPATHTOF(data\Keycard\UI\gear_Keycard_C1_CA.paa); \
+		}; \
+		class Level2 { \
+			name    = CSTRING(DoorState_level2); \
+			tooltip = CSTRING(DoorState_level2_tooltip); \
+			value = 2; \
+			picture = QPATHTOF(data\Keycard\UI\gear_Keycard_C2_CA.paa); \
+		}; \
+	}; \
+};
 
-#define NORMAL_GLASS_HITPOINT(glassID,arm,rad) \
-			class Glass_##glassID##_hitpoint \
-			{ \
-				armor = arm; \
-				material = -1; \
-				name = Glass_##glassID; \
-				visual = Glass_##glassID##_hide; \
-				passThrough = 0; \
-				radius = rad; \
-				convexComponent = Glass_##glassID##_hide; \
-				class DestructionEffects \
-				{ \
-					class BrokenGlass1 \
-					{ \
-						simulation = "particles"; \
-						type = "BrokenGlass1NN"; \
-						position = Glass_##glassID##_effects; \
-						intensity = 0.15000001; \
-						interval = 1; \
-						lifeTime = 0.05; \
-					}; \
-					class BrokenGlass2: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass2NN"; \
-					}; \
-					class BrokenGlass3: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass3NN"; \
-					}; \
-					class BrokenGlass4: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass4NN"; \
-					}; \
-					class BrokenGlass5: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass5NN"; \
-					}; \
-					class BrokenGlass6: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass6NN"; \
-					}; \
-					class BrokenGlass7: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass7NN"; \
-					}; \
-					class BrokenGlass1S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass1SN"; \
-					}; \
-					class BrokenGlass2S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass2SN"; \
-					}; \
-					class BrokenGlass3S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass3SN"; \
-					}; \
-					class BrokenGlass4S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass4SN"; \
-					}; \
-					class BrokenGlass5S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass5SN"; \
-					}; \
-					class BrokenGlass6S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass6SN"; \
-					}; \
-					class BrokenGlass7S: BrokenGlass1 \
-					{ \
-						type = "BrokenGlass7SN"; \
-					}; \
-				}; \
-			};
+#define ATTRIBUTES_BUTTON_EXPRESSION \
+class PressExpression \
+{ \
+	displayName     = CSTRING(PressExpression_displayname); \
+	tooltip         = CSTRING(PressExpression_tooltip); \
+	property        = "PressExpression"; \
+	control         = "EditCodeMulti5"; \
+	expression      = "_this setVariable ['SCP_CB_Operetable_PressExpression', _value];"; \
+	defaultValue    = """[[myDoor0, myDoor1], 'None'] call SCP_fnc_buttonTrigger;"""; \
+	typeName        = "STRING"; \
+};
+
+#define ATTRIBUTES_CARDREADER_REQUIRED_CLEARANCE_LEVEL \
+class RequiredClearanceLevel \
+{ \
+	displayName     = CSTRING(RequiredClearanceLevel_displayname); \
+	tooltip         = CSTRING(RequiredClearanceLevel_tooltip); \
+	property        = "RequiredClearanceLevel"; \
+	control         = "Combo"; \
+	expression      = "_this setVariable ['SCP_CB_Operetable_Required_ClearanceLevel', _value];"; \
+	defaultValue    = 1; \
+	typeName        = "NUMBER"; \
+	class Values { \
+		class Level1 { \
+			name    = CSTRING(RequiredClearanceLevel_level1); \
+			tooltip = CSTRING(RequiredClearanceLevel_level1_tooltip); \
+			value   = 1; \
+			picture = QPATHTOF(data\Keycard\UI\gear_Keycard_C1_CA.paa); \
+		}; \
+		class Level2 { \
+			name    = CSTRING(RequiredClearanceLevel_level2); \
+			tooltip = CSTRING(RequiredClearanceLevel_level2_tooltip); \
+			value = 2; \
+			picture = QPATHTOF(data\Keycard\UI\gear_Keycard_C2_CA.paa); \
+		}; \
+		class Level3 { \
+			name    = CSTRING(RequiredClearanceLevel_level3); \
+			tooltip = CSTRING(RequiredClearanceLevel_level3_tooltip); \
+			value = 3; \
+			picture = QPATHTOF(data\Keycard\UI\gear_Keycard_C3_CA.paa); \
+		}; \
+		class Level4 { \
+			name    = CSTRING(RequiredClearanceLevel_level4); \
+			tooltip = CSTRING(RequiredClearanceLevel_level4_tooltip); \
+			value = 4; \
+			picture = QPATHTOF(data\Keycard\UI\gear_Keycard_C4_CA.paa); \
+		}; \
+		class Level5 { \
+			name    = CSTRING(RequiredClearanceLevel_level5); \
+			tooltip = CSTRING(RequiredClearanceLevel_level5_tooltip); \
+			value = 5; \
+			picture = QPATHTOF(data\Keycard\UI\gear_Keycard_C5_CA.paa); \
+		}; \
+	}; \
+};
