+++ addons/room_things/functions/fnc_doorOperate.sqf
@@ -123,9 +123,9 @@ switch (_operationType) do {
 // Randomly selects a sound from _soundSet to play. 
 _randomSound = _soundSet select floor random count _soundSet;
 
-// Plays _randomSound from _object for door on every client but server.
-//[_soundOrigin, _randomSound] remoteExec ["say3D", 0];
-_soundOrigin say3D _randomSound;
+// Plays _randomSound from _object for door.
+[_soundOrigin, _randomSound] remoteExec ["say3D", 0];
+//_soundOrigin say3D _randomSound;
 
 //Removes _soundOrigin after _soundDuration seconds.
 [_soundOrigin, _soundDuration] spawn {
